pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "treasure"
version = "0.0.1"
requires-python = ">=3.6"
dependencies = [
    "art ~= 5.7",
    "PyYAML ~= 6.0"
]

[tool.setuptools.packages.find]
include = ["treasure"]  # package names should match these glob patterns

[project.optional-dependencies]
dev = [
    "black ~= 21.12b0",
    "flake8 ~= 4.0.1",
    "isort ~= 5.10.1",
    "pytest ~= 7.1.2"
]

[project.scripts]
greet = "treasure.cli:main"
